+++ packages/builder/examples/app.ts
@@ -1,5 +1,7 @@
 import { App } from '@backan/core'
-export const app = new App( {
+
+type AppEnv = {DB: object}
+export const app = new App<AppEnv>( {
 	version     : '1.0.0',
 	title       : 'BACKAN Example app',
 	description : 'API documentation for BACKAN Example',