+++ packages/builder/examples/md.ts
@@ -1,6 +1,7 @@
 import { buildMD } from '../src/md'
 import { app }     from './app'
+
 buildMD( {
 	app    : app,
-	output : 'example-openapi.md',
+	output : 'openapi.md',
 } )