+++ packages/builder/examples/schema.ts
@@ -1,6 +1,7 @@
 import { buildSchema } from '../src/schema'
 import { app }         from './app'
+
 buildSchema( {
-	app    : app,
-	output : 'example-openapi.json',
+	app,
+	output : 'openapi.json',
 } )